refactor(perf): split src/pipeline/stages/autofix.ts — extract autofix-agent.ts
